fix(ci): split render.rs under the size guard, refresh llms.txt
Three failures on #752, one of them structural.

**`stella-tui/src/render.rs` crossed the 1500-line guard at 1514.** It is not
grandfathered, so a baseline entry is not an option — that is the rule that
stops the tree getting worse. Split along the concern rather than the line
count: everything that turns model state into `Line`s moved to
`render/entry.rs` (747 lines), everything that draws into a `Frame`/`Buffer`
stayed (800 lines). `render/row.rs` is unchanged and remains the word layer
beneath — `entry_lines` composes the sentence, `row` renders the phrase.

Verified as a **pure move**: the extracted body diffs byte-identical against
the original range. `render` re-exports the two `pub(crate)` entry points, so
`ui.rs` and `deck_ui.rs` call sites are untouched. `resolve_inline_diff` and
`INLINE_DIFF_CAP` stayed with the draw side and are reached through `super::`,
which needed no visibility change. The leftover imports moved with their users;
`render/tests.rs` now names the three it had been getting through
`use super::*`.

**`deck_render.rs` (+28) and `model.rs` (+42)** are grandfathered and their
growth is the transcript work itself, so their baselines are updated — the
sanctioned path for a file already over the ceiling, and the diff makes the
increase visible in review.

**`docs/llms.txt` regenerated** — the docs changes here landed without it.

Also merges `origin/main`, which had moved twice under this branch (#750, and
Phase 4 in #751); the merge is clean and the baseline file reconciles.

`make gate` green: 3868 tests, clippy clean, no file over its ceiling.
