+++ scripts/file-size-baseline.txt
@@ -28,7 +28,7 @@
 1556 stella-tools/src/media.rs
 3014 stella-tools/src/registry.rs
 1797 stella-tools/src/scripts.rs
-1677 stella-tui/src/deck_render.rs
+1705 stella-tui/src/deck_render.rs
 4049 stella-tui/src/deck_ui.rs
-1609 stella-tui/src/model.rs
+1651 stella-tui/src/model.rs
 1652 stella-tui/src/views/engine.rs
